Refactor: tools,libpacemaker: Create new top-level cluster status messages.
These are the old print_* functions from crm_mon, but turned into
formatted output messages.
