+++ tools/crm_mon_curses.c
@@ -15,6 +15,7 @@
 #include <crm/fencing/internal.h>
 #include <crm/pengine/internal.h>
 #include <glib.h>
+#include <pcmki/pcmki_output.h>
 
 #include "crm_mon.h"
 
@@ -428,6 +429,19 @@ cluster_maint_mode_console(pcmk__output_t *out, va_list args) {
     }
 }
 
+PCMK__OUTPUT_ARGS("cluster-status", "pe_working_set_t *", "crm_exit_t",
+                  "stonith_history_t *", "gboolean", "unsigned int",
+                  "unsigned int", "const char *", "GList *", "GList *")
+static int
+cluster_status_console(pcmk__output_t *out, va_list args) {
+    int rc = pcmk_rc_no_output;
+
+    blank_screen();
+    rc = pcmk__cluster_status_text(out, args);
+    refresh();
+    return rc;
+}
+
 PCMK__OUTPUT_ARGS("stonith-event", "stonith_history_t *", "gboolean", "gboolean")
 static int
 stonith_event_console(pcmk__output_t *out, va_list args) {
@@ -479,6 +493,7 @@ static pcmk__message_entry_t fmt_functions[] = {
     { "cluster-dc", "console", pe__cluster_dc_text },
     { "cluster-options", "console", pe__cluster_options_text },
     { "cluster-stack", "console", pe__cluster_stack_text },
+    { "cluster-status", "console", cluster_status_console },
     { "cluster-summary", "console", pe__cluster_summary },
     { "cluster-times", "console", pe__cluster_times_text },
     { "failed-action", "console", pe__failed_action_text },
@@ -543,3 +558,18 @@ crm_mon_register_messages(pcmk__output_t *out) {
 }
 
 #endif
+
+void
+blank_screen(void)
+{
+#if CURSES_ENABLED
+    int lpc = 0;
+
+    for (lpc = 0; lpc < LINES; lpc++) {
+        move(lpc, 0);
+        clrtoeol();
+    }
+    move(0, 0);
+    refresh();
+#endif
+}
